Add payment processing components: implement Payment service, controller, and DTOs; configure Kafka integration for payment notifications

+++ order-service/src/main/java/com/mlue/ecommerce/dto/OrderConfirmationDto.java
@@ -3,13 +3,12 @@ package com.mlue.ecommerce.dto;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.mlue.ecommerce.constant.PaymentMethod;
-import java.math.BigDecimal;
 import java.util.List;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record OrderConfirmationDto(
         String orderReference,
-        BigDecimal totalAmount,
+        double totalAmount,
         PaymentMethod paymentMethod,
         CustomerResponseDto customer,
         List<ProductPurchaseResponseDto> products
