+++ order-service/src/main/java/com/mlue/ecommerce/dto/OrderDto.java
@@ -7,8 +7,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-
-import java.math.BigDecimal;
 import java.util.List;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
@@ -16,7 +14,7 @@ public record OrderDto(
         String reference,
 
         @Positive(message = "Amount must be positive")
-        BigDecimal totalAmount,
+        double totalAmount,
 
         @NotNull(message = "Payment method is required")
         PaymentMethod paymentMethod,
